Kirby Starterkit 2.4.0

+++ panel/app/src/panel/models/file/menu.php
@@ -56,7 +56,10 @@ class Menu {
 
     $list->append($this->previewOption());
     $list->append($this->editOption());
-    $list->append($this->deleteOption());
+
+    if($this->file->ui()->delete()) {      
+      $list->append($this->deleteOption());
+    }
 
     return '<nav class="dropdown dropdown-dark contextmenu">' . $list . '</nav>';
